separate documentation tools in ethereum_spec_tools
