+++ setup.cfg
@@ -13,6 +13,7 @@ classifiers =
 [options]
 packages =
     ethereum_spec_tools
+    ethereum_spec_tools/doc
     ethereum_spec_tools/lint
     ethereum_spec_tools/lint/lints
     ethereum
@@ -99,7 +100,7 @@ ethereum =
 console_scripts =
     ethereum-spec-lint = ethereum_spec_tools.lint:main
     ethereum-spec-sync = ethereum_spec_tools.sync:main
-    ethereum-spec-diff = ethereum_spec_tools.diff:main
+    ethereum-spec-diff = ethereum_spec_tools.doc.diff:main
     ethereum-spec-new-fork = ethereum_spec_tools.new_fork:main
 
 [options.extras_require]
