+++ src/ethereum_spec_tools/doc/__init__.py
@@ -1,0 +1,6 @@
+"""
+Documentation Tools
+^^^^^^^^^^^^^^^^^^^
+
+Tools used to document the specifications
+"""
