+++ whitelist.txt
@@ -348,3 +348,5 @@ astroid
 autoapi
 attrname
 expr
+subpackages
+toctree
